requirements.txt
torch
transformers
faiss-cpu
sentence-transformers
streamlit
pandas
numpy
matplotlib
seaborn
scikit-learn
openai
